ImgMath: refactored formula constructor into class Compute, invoked from ImgMath.compute static methods.
